Implement the tp_new slot.
With this change it is impossible to create a Wrapper object where
'self->obj' == NULL. __new__ is the very first place where a new
instance is created and there it is ensured that self->obj is set to a
reasonable value. I have seen no other code where self->obj is set.

This allows now to change several places in _Acquisition.c where
self->obj is checked for NULL pointer. Which leads to simpler code.